fix: repair CI workflows and qualify release locally (v2.69)
Fix environment-specific CI failures and produce local release
qualification artifacts. Phases 284-286 complete on disk; hosted
GitHub Actions rerun required before tagging release candidate.

CI workflow repairs:
- Fixed MSRV compatibility issues in ci.yml
- Updated release-qualification.yml for SDK parity and conformance waves
- Hardened flaky test assumptions discovered during tarpaulin coverage runs
  (web3_e2e_qualification, receipt_query, mcp_serve_http, remote_mcp,
  trust_control, runtime_devnet)

Source fixes discovered during CI/coverage runs:
- Fixed test precondition checks across arc-settle, arc-anchor, arc-cli
- Hardened status assumptions in receipt query and MCP serve HTTP tests
- Fixed timeout/initialization issues in remote MCP and trust control
- Updated monetary type tests, credential discovery, and policy matchers

Release qualification:
- All 5 conformance waves proven locally (JS, Python, Go peers)
- Signed per-wave certification artifacts generated via qualify-release.sh
- Qualification bundle at target/release-qualification/artifact-manifest.json
- Coverage gate integrated (COVERAGE_FAIL_UNDER=67)

Blocker: hosted GitHub Actions must run green on current main before
CI-01 through CI-05 can be marked complete and release candidate tagged.

+++ crates/arc-settle/src/payments.rs
@@ -143,7 +143,7 @@ pub fn build_x402_payment_requirements(
             .capability_id
             .clone()
             .unwrap_or_else(|| dispatch.dispatch_id.clone()),
-        amount_minor_units: u64::from(dispatch.settlement_amount.units),
+        amount_minor_units: dispatch.settlement_amount.units,
         currency: dispatch.settlement_amount.currency.clone(),
         settlement_mode,
         governed_authorization_required: true,
